backend: move logic into app directory and split into modules
This moves the core flask app logic into a subdirectory and splits the
different concerns into their own modules.

Application initialization occurs in the __init__.py. Here we create the
database instance that will be used by the application, this needs to
happen after the flask application is created. We also import the models
and routes modules at the bottom.

This commit also deletes the old app.py file.