frontend: general local storage handling
Check if a cached game_id exists, if so then set that to be the game_id
used. When user clicks New Game clear local storage but persist the
build_time.

In general this commit finishes implementing the ability for a user to
refresh the page without clearing current progress on the game.